refactor: Simplified the folder structure
