--- packages/ember-codemod-ember-render-helpers-to-v1/src/migration/ember-app/steps/index.ts
@@ -1,1 +1,0 @@
-export * from './update-templates.js';
