--- packages/ember-codemod-ember-render-helpers-to-v1/src/migration/ember-v1-addon/index.ts
@@ -1,9 +1,0 @@
-import type { CodemodOptions } from '../../types/index.js';
-import { createOptions } from '../../utils/steps/create-options.js';
-import { updateTemplates } from './steps/index.js';
-
-export function migrateEmberV1Addon(codemodOptions: CodemodOptions): void {
-  const options = createOptions(codemodOptions);
-
-  updateTemplates(options);
-}
