--- packages/ember-codemod-ember-render-helpers-to-v1/src/migration/ember-v1-addon/steps/index.ts
@@ -1,1 +1,0 @@
-export * from './update-templates.js';
