--- packages/ember-codemod-ember-render-helpers-to-v1/src/migration/index.ts
@@ -1,2 +1,0 @@
-export * from './ember-app/index.js';
-export * from './ember-v1-addon/index.js';
