+++ packages/ember-codemod-ember-render-helpers-to-v1/tests/index/my-app.test.ts
@@ -1,0 +1,18 @@
+import { assertFixture, loadFixture, test } from '@codemod-utils/tests';
+
+import { runCodemod } from '../../src/index.js';
+import { inputProject, outputProject } from '../fixtures/my-app/index.js';
+import { codemodOptions } from '../helpers/shared-test-setups/my-app.js';
+
+test('index > my-app', function () {
+  loadFixture(inputProject, codemodOptions);
+
+  runCodemod(codemodOptions);
+
+  assertFixture(outputProject, codemodOptions);
+
+  // Check idempotence
+  runCodemod(codemodOptions);
+
+  assertFixture(outputProject, codemodOptions);
+});
