+++ packages/ember-codemod-ember-render-helpers-to-v1/tests/index/my-v1-addon.test.ts
@@ -1,0 +1,18 @@
+import { assertFixture, loadFixture, test } from '@codemod-utils/tests';
+
+import { runCodemod } from '../../src/index.js';
+import { inputProject, outputProject } from '../fixtures/my-v1-addon/index.js';
+import { codemodOptions } from '../helpers/shared-test-setups/my-v1-addon.js';
+
+test('index > my-v1-addon', function () {
+  loadFixture(inputProject, codemodOptions);
+
+  runCodemod(codemodOptions);
+
+  assertFixture(outputProject, codemodOptions);
+
+  // Check idempotence
+  runCodemod(codemodOptions);
+
+  assertFixture(outputProject, codemodOptions);
+});
